Armorer's Ritual bargain sale
Seeing as how this is is quite literally a staple of SS13's heretic, literally being on the icon for the antagonist itself, why the fuck is this restricted to just rust/blade heretics? Well, to keep some uniqueness, non-rust/blade heretics may pick up this side knowledge for 2 points rather than 1. This won't apply to heretic paths who have void cloak as a side path (Void for instance.)