New Heretic DLC: Time Path (First pass)
After a LOT of work, this is in a state were i'd allow it to be officially commited. Time Heretics are a path revolving around Time Manipulation and AGE, specializing in evasion and recovery. A real hit-and-runner path. They've little in the way of offense or defense. Soon i'll have another path added (Nature) which will sit right next to it. Currently this path is sitting between Cosmos and Ash. Soon it'll be Cosmos and Nature.