Some Syndicate MODsuit changes.
Giving them more module room to account for the new modules added to the uplink, ain't running out of complexity just the worst? Also some improvements to the dark paladin MODsuit available in the uplink, and makes the interdyne MODsuit in the uplink much less of a hassle: Now it comes with the standard modules you'd expect: DNA lock, jetpack, shock absorber, emp shield, pathfinder, and even a magnetic harness.