Ports Heliostation straight from Fulp.
Took me about a day of on-and-off work, but i think this is in a playable state now. The AI sat's solars aren't functional though, and due to my lack of knowledge over how solar's code works... well. Let's just say a patch will come later when i figure that out.