Better Freedom Implant.
This one's barely ever used, except on MRP. Most secoffs just continuously stun you or beat you into crit regardless of whether or not you're cuffed. It's really only good against bolas, but that's a stretch. Admittedly, the worst part about this is that it's 5 TC, which is one quarter of most tot's TC, and it only has 4 uses. It's def not worth 5 TC, more like 2. Regardless, this implant still had potential if used right and i had this idea in mind for this implant and two others. (The freedom implant now regains it's charges over time. Every 3 minutes = 1 charge)