The anomaly belt! (Requested by Nexus)
An alternate take on the reactive armor, this belt can be slotted with any anomaly core to gain a unique ability to be used by it's wearer. The abilities are ACTIVE not PASSIVE, you don't need to be hit like the reactive armor for these effects to proc. (I won't list each effect here, but i will say most of the belt effects are similar to the anomalies from which they come from.)