Allows roundstart traitors to buy a nukie-team implant during warops
Also improves the Operative Induction kit by giving it a syndicate MODsuit (and combat gloves) as opposed to a spare space suit. Warops (and nukies in general) are well-known for not collaborating with traitors on station and the whole POINT of doing so is basically moot. So let's give some more incentive to actually use this cheap reinforcement option and allow traitors to opt-in without having to directly meet the nukie team themselves at the cost of any additional gear (they'll like sabotage the crew anyways, let's not buff them any further)