Chemi-compiler buffs
Overall, while being able to do chem wherever you want, the regen is slow and the max capacity was 50 PER REAGENT. It was powerful if you needed medicine on the fly, but for any major chemistry work to be done it was far cheaper to just invade chemistry and get whatever you need done. I'm dropping the cost to 8 TC, increasing the maximum reagent capacity to 200 and increasing the reagent generation from 1 per second to 5 per second. Should be more reliable for on-the-go chemistry, and less punishing if you hit the wrong button.