Syndicate Surgery Duffel Bag Buff + Self-Surgery Implant
Added a special implant to the Syndicate Surgery Duffel Bag which allows the implantee to perform self-surgery. Now this item is much-less of a niche and can actually be rather useful -- especially since you no longer require an ally to constantly be there whenever you need surgery. The item now costs 4 TC instead of 3 TC to compensate.