Boosts the vending credits collection amount from 0.2 to 0.5
Originally whenever an item is purchased, 20% of the credits used to buy that item is stored in the vending machine which can be reclaimed by restocking that vendor or by breaking it/fishing the money out. Additionally, 50% of that 20% (rounded up) is immediately given to cargo as additional funding. Now that 20% is 50%. So if you bought a medkit for 500 credits, if you immediately restock the machine, you'd get 250 credits back (instead of 100 credits) and cargo would receive 125 credits (instead of 50 credits). I'm boosting this value because nobody ever restocks vendors and i don't think most people even know you can get money out of doing so, and even if they did, a 20% credit return on both finding a vending machine with money inside it (can be revealed by examining it), obtaining the appropriate vending restock, and running the restock to said vendor for a restock is hardly any incentive. Obviously, it's supposed to be cargo's job to restock vendors, but nobody does it anyways. Too much time for too little money, ya know?