Ports Monkiestation 2.0's Cortical Borers
Cortical borers got removed from TG a while back and while i never got to see them in action, I absolutely loved the idea of em. Cortical borers are a midround antagonist (or a 20 TC tot purchase) that burrow into the head of another to support their host or leech off them all in order to keep their species alive and thriving. -- After much grueling on-and-off work i believe i've gotten this thing in... passable order, and will be on the lookout for any bugs.