Adds several new security alert levels:
Yellow, Amber, Gamma, Lambda, and Epsilon respectively. Yellow can be issues from a communication console, Amber is currently auto-issued after a Blob Announcement is made, Gamma is admin only, Lambda occurs when nar'sie is summoned successfully, and Epsilon is either given should the crew piss off CC or when a revolution is successful.