Revolution Changes
A revolutionaries victory within a round no longer results in the round ending. Instead, (if the shuttle can be called) the shuttle is called, and the ability to cancel the shuttle call is disabled. Additionally, since a rev win results in a Epsilon alert that cannot be removed, the shuttle timer is around 5x as long, during which a deathsquad will (attempt) to spawn with the mission of exterminating everyone on the station (or setting off the self-destruct nuke). Just to add salt onto the wound: FIFTY threat is added to the midround budget after a rev round to ensure some more midrounds roll to keep the round fresh, just in case.