Ports Monkiestation 2.0's Clockwork Cult
After a LOT of work, i've got much of this antag ported and functional. Some things are still a tad broken, like the TGUI, of which i still intend to fix later (I'm not even remotely adept with tgui so that may take some time) but most of it is completely functional save for some progress bars not having the icon/text over them and some missing colors. I'm releasing this for now so i can start playtesting with other people and (POSSIBLY? MAYBE?) have some help with the tgui fixes.