Gives clockies two unique items for saccing culties on a vitality matrix
The most likely item to receive is the Vitality Clockwork Slab, a direct upgrade to the regular slab that grants passive regeneration to the user should it be held or pocketed. The second possible item is the shielded clockwork armor, the clockies version of the blood cult's shielded robes which can be upgraded at a stargazer for better armor.