The Regenerate Implant makes it's debut
A little pet peeve of mine i had a while back. The changeling extract medipen was nice and all, but if you lost both arms you were still screwed. Now there's an implant which does it all for ease of use for 4 TC. I'll lower the changeling extract medipen to 1 TC to not make the implant the default option. It starts off with no charges and takes 10 minutes regain 1 charge to avoid on-demand application (that's what the medipen is for) for a maximum of 2 charges. Additionally, to help make the ling extract even better aside from losing limbs/decayed organs they'll also remove certain wounds.