Returns Cryogelidia's functionality pre-bug and pre-"fixes"
So, a while back cryogelidia got broken, and people tried to fix it. They succeeded to an extent, with purified cryogelidia being to only way to actually use the inverted chem. They also removed the ability to use the chem via ingesting for some fucking reason? This returns the ability to use cryogel in a non-purified state (just like every other inverted chem) and re-enables ingested effects for cryogel.