Adds two inverse chemicals (and gives some chem recipes ph requirements)
Adds Razorium as an inverse of bicaridine, which deals brute damage over time and when applied via touch will apply additional brute damage, razorium foam bombs can be lethal. Adds Insuzine as an inverse of Kelotane, which heals burns twice as much as kelotane but drops the patient's temperature and has a low overdose requirement (12), additionally while overdosing the chemical will freeze the patient in a protective block of ice should they be cold. Insuzine, for it's namesake purges most electrical-based reagents quickly.