Another New Traitor Objective.
Traitors now have a small chance of receiving an "Assist" objective. This requires their target to be an antagonist and complete all of their objectives. There's an 80% chance the target is a traitor, and a 20% chance for any possible kind of antagonist. (Of course, there could be no other antagonists/traitors and you'll get a false positive that'll get you busted.)