Gives the energy sword and dagger the ability to reflect projectiles.
The sword is given a flat 50% to reflect a projectile rather than outright block it, and the dagger is given a 20% chance to reflect projectiles rather than block... and a 20% to block things in general. I'm mostly adding this to make the energy dagger stand out far more from the much more cost effective cybersun pen and actually give it decent stats to boot. I did the same to the energy sword just because i could (Energy swords reflect in ss14 and thought, why does the dual saber reflect but not the regular one?)