Flagellant's robes changes
No longer provides massive negative armor, instead providing VERY low armor values, and reduces the 30% speed up for the armor/hood to 20% for a maximum of 40% movement speed boost. I completely agree with a recent removal PR of these robes that they're a bit of a noob trap, i decided to balance them to be less annoying to fight and less detrimental to those who wear it.