Adds the Liquid Anomaly (and it's core-based items)
The Liquid Anomaly can now spawn on station, and every twenty seconds it'll spill 100 units of a random reagent determined upon it's creation directly onto the floor. This can cause certain rooms to become hazardous to life (drowning) or produce large amounts of both harmful and helpful reagents which may or may not be impactful during a round. Should it go critical, it'll spawn reagent elementals that'll attack nearby people, when the elements die and/or are gibbed/butchered they'll spill a random reagent onto the floor determined upon their creation.