Adds the Anomaly Stabilizer -- a new way to get research points!
Adds a sort of inverse to the Anomaly Neutralizer and a much simpler take on the anomaly research tech seen in SS14. After acquiring anomaly technology, you may print out an anomaly stabilizer to stabilize an anomaly, pausing it's decay timer and making it grant 1 research point every second. You may still neutralize it for it's core.