Adds the Power Exporter as an exclusive CE machine.
Reworked idea from this PR: https://github.com/tgstation/tgstation/pull/24564

This single machine when build on the powernet acts almost like a power sink, but only when the powernet has a surplus amount of energy to draw from. The excess power is lost, but in return the engineering department gains roughly 70% of the Kilowatts sent as Engi dept credits, with 30% sent straight to the cargo budget. Starts off with a -0.75% penalty with tier 1 parts, up to the full 100% payout with tier 4 parts.

Finally! An actual reason for the engineers to powermax the SM, set up solars, or learn that other atmos power machine i can't even name because nobody fucking touches the damn thing.