Adds a Bank Heist secondary objective for traitors
It functions quite like the "hack communications console" objective in which traitors must compete to complete it first. Completion requires the traitor to infiltrate the vault on station and hack a bank machine within it. If successful, the account in the bank machine is drained completely of funds and is put in debt by a thousand or so (default account is cargo budget).