Adds a secondary objective for traitors: Syndicate Monkey Recruitment
Asks a traitor to deliver a monkey previously enslaved to the whims of Nanotrasen employees to an extraction pod in a specific area, which is then indoctrinated into a Syndicate Monkey (ghost role), which will then be sent back to the station 6 minutes later to cause havoc. These Monkeys are NOT enslaved to the traitor who initially sent them. Monkey is now free monkey.