Adds Commando Operatives as a potential variant of Nukies
Ported from Monkiestation 2.0. The PR can be found here: https://github.com/Monkestation/Monkestation2.0/pull/6824  This version of nukies starts off with 0 TC, but a boatload of starting gear and a syndicate requisitions vendor. Commando Ops are expected to bring a destructable nuclear warhead over to the station and defend it for some time until it blows up.